requirements.txt
# requirements.txt
requests>=2.26.0
pandas>=1.3.0
python-dotenv>=0.19.0
tqdm>=4.62.0